Applied hoisting technique
compress-ctxtdb.cpp / decompress-ctxtdb.sage is for the case of ci-
phertext database.
compress-ptxtdb.cpp / decompress-ptxtdb.sage is for the case of pl-
aintext database.
compress.cpp and decompress.sage is the ciphertext database case,
which our current paper mostly focusing on.

We also applied the hoisting technique for ctxtdb case. That is, we
removed some redundant keyswwitchings during Pegasus procedure.